Introduce Graphiti::Serializer
The main intent is to avoid specifying
`JSONAPI::Serializable::Resource`, which is a confusing superclass. But
this also gives us a place to stuff overrides, and mostly paves the way
for avoiding conflicts with raw usage.
